refactor: transfer mock-server to setUpTests
