+++ react-components/setupTests.js
@@ -1,8 +1,23 @@
 import { expect, afterEach } from 'vitest';
 import { cleanup } from '@testing-library/react';
 import matchers from '@testing-library/jest-dom/matchers';
+import { rest } from 'msw';
+import { setupServer } from 'msw/node';
+import { mock } from './src/__tests__/mockData';
 
 expect.extend(matchers);
+
+const server = setupServer(
+  rest.get('https://rickandmortyapi.com/api/character', (req, res, ctx) => {
+    return res(ctx.json(mock));
+  })
+);
+
+beforeAll(() => server.listen());
 afterEach(() => {
+  server.resetHandlers();
   cleanup();
+});
+afterAll(() => {
+  server.close();
 });
