+++ react-components/src/__tests__/MainPage.test.tsx
@@ -1,21 +1,8 @@
 import '../../setupTests';
 import { render, screen, waitFor } from '@testing-library/react';
 import { MainPage } from '../pages';
-import { rest } from 'msw';
-import { setupServer } from 'msw/node';
+import '../../setupTests';
 import { mock } from './mockData';
-
-const server = setupServer(
-  rest.get('https://rickandmortyapi.com/api/character', (req, res, ctx) => {
-    return res(ctx.json(mock));
-  })
-);
-
-beforeAll(() => server.listen());
-afterEach(() => server.resetHandlers());
-afterEach(() => {
-  server.close();
-});
 
 describe('Tests for MainPage', () => {
   test('data from API are render on MainPage', async () => {
